requirements.txt
Flask
Flask-Cors
Flask-Swagger
Flask-JWT-Extended
bcrypt
sqlalchemy
tensorflow
Pillow
nbformat
nbconvert